+++ src/providers/extractors/animeextractor.js
@@ -61,15 +61,15 @@ export default class Extractor extends BaseExtractor {
     let slug = animeTitle.replace(/[!]/gi, "").replace(/\s-\s/gi, "").replace(/\s+/g, "-").toLowerCase();
     slug = slug in animeMap ? animeMap[slug] : slug;
 
-    let season, episode, quality;
+    const quality = torrent.title.match(/(\d{3,4})p/) !== null ? torrent.title.match(/(\d{3,4})p/)[0] : "480p";
+
+    let season, episode;
     if (torrent.title.match(regex).length >= 5) {
       season = parseInt(torrent.title.match(regex)[2], 10);
       episode = parseInt(torrent.title.match(regex)[3], 10);
-      quality = torrent.title.match(regex)[4];
     } else {
       season = 1;
       episode = parseInt(torrent.title.match(regex)[2], 10);
-      quality = torrent.title.match(regex)[3];
     }
 
     const episodeTorrent = {
@@ -106,12 +106,15 @@ export default class Extractor extends BaseExtractor {
     const secondSeason = /\[horriblesubs\].(.*).S(\d)...(\d{2,3}).\[(\d{3,4}p)\]/i;
     const oneSeason = /\[horriblesubs\].(.*)...(\d{2,3}).\[(\d{3,4}p)\]/i;
     const animerg = /\[animerg\]\s+(\D+)\s\-\s(\d{3}|\d{2})\D+(\d{3,4}p)/i;
+    const commie = /\[Commie\].(\D+)...(\d{2,3}).*\.mkv/i;
     if (torrent.title.match(secondSeason)) {
       return this._extractAnime(torrent, secondSeason);
     } else if  (torrent.title.match(oneSeason)) {
       return this._extractAnime(torrent, oneSeason);
     } else if (torrent.title.match(animerg)) {
       return this._extractAnime(torrent, animerg);
+    } else if (torrent.title.match(commie)) {
+      return this._extractAnime(torrent, commie);
     } else {
       logger.warn(`${this.name}: Could not find data from torrent: '${torrent.title}'`);
     }
@@ -172,6 +175,7 @@ export default class Extractor extends BaseExtractor {
       logger.info(`${this.name}: Total pages ${totalPages}`);
 
       const torrents = await this._getAllTorrents(totalPages, provider);
+      console.log(torrents.length);
       const animes = await this._getAllAnimes(torrents);
       return await asyncq.mapLimit(animes, maxWebRequest, anime => this.getAnime(anime));
     } catch (err) {